requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Cython==0.29.32
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
emoji==1.2.0
filelock==3.8.0
flatbuffers==22.10.26
gast==0.4.0
gluonnlp==0.10.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kss==3.6.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
more-itertools==9.0.0
mxnet==1.9.1
networkx==2.8.8
nltk==3.7
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pyparsing==3.0.9
python-mecab-kor==1.2.3
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
